+++ src/audio/CMakeLists.txt
@@ -185,6 +185,7 @@ if(CONFIG_IPC_MAJOR_3)
 	set(tdfb_sources multiband_drc/multiband_drc_ipc3.c)
 	set(dcblock_sources dcblock/dcblock_ipc3.c)
 	set(mux_sources mux/mux_ipc3.c)
+	set(crossover_sources crossover/crossover_ipc3.c)
 elseif(CONFIG_IPC_MAJOR_4)
 	set(volume_sources volume/volume.c volume/volume_generic.c volume/volume_ipc4.c)
 	set(asrc_sources asrc/asrc_ipc4.c)
@@ -195,6 +196,7 @@ elseif(CONFIG_IPC_MAJOR_4)
 	set(tdfb_sources multiband_drc/multiband_drc_ipc4.c)
 	set(dcblock_sources dcblock/dcblock_ipc4.c)
 	set(mux_sources mux/mux_ipc4.c)
+	set(crossover_sources crossover/crossover_ipc4.c)
 endif()
 set(mixer_sources ${mixer_src})
 set(asrc_sources asrc/asrc.c asrc/asrc_farrow.c asrc/asrc_farrow_generic.c)
@@ -204,7 +206,7 @@ set(dcblock_sources dcblock/dcblock.c dcblock/dcblock_generic.c dcblock/dcblock_
 set(crossover_sources crossover/crossover.c crossover/crossover_generic.c)
 set(tdfb_sources tdfb/tdfb.c tdfb/tdfb_generic.c tdfb/tdfb_direction.c)
 set(drc_sources drc/drc.c drc/drc_generic.c drc/drc_math_generic.c)
-set(multiband_drc_sources multiband_drc/multiband_drc_generic.c crossover/crossover.c crossover/crossover_generic.c drc/drc.c drc/drc_generic.c drc/drc_math_generic.c multiband_drc/multiband_drc.c )
+set(multiband_drc_sources multiband_drc/multiband_drc_generic.c crossover/crossover.c drc/drc.c drc/drc_generic.c drc/drc_math_generic.c multiband_drc/multiband_drc.c )
 set(mfcc_sources mfcc/mfcc.c mfcc/mfcc_setup.c mfcc/mfcc_common.c mfcc/mfcc_generic.c mfcc/mfcc_hifi4.c mfcc/mfcc_hifi3.c)
 set(mux_sources mux/mux.c mux/mux_generic.c)
